refactor(browser): split workspace_manager.go into focused modules for better maintainability
